Add filmstrip slide
This is broken in Chrome because <use> in Chrome is just buggy as
anything.

The fact that it animates is not the primary issue but rather that it
will apply the `transform` applied to a <use>-ed element but NOT the
`transform-origin` or `transform-box`... unless you touch the element in
DevTools -- then it starts working as it should.